Factor out MLIR-to-flatbuffer and get-lce-target functions into separate file
